+++ packages/openapi/src/config-extensions/examples.ts
@@ -1,0 +1,10 @@
+export interface Example<T = any> {
+  example?: T;
+  examples?: Record<
+    string,
+    {
+      value: T;
+      summary?: string;
+    }
+  >;
+}