+++ packages/swagger/src/index.ts
@@ -5,4 +5,4 @@ export type { Options } from './types';
 export { swagger };
 export default swagger;
 
-export { defineEndpoint, EndpointDocumentation } from './openapi';
+export { defineEndpoint, EndpointDocumentation, Example } from './openapi';